Avoid the C-style cast

+++ distbench_engine.cc
@@ -894,8 +894,8 @@ void DistBenchEngine::RunAction(ActionState* action_state) {
         period + absl::UnixEpoch() + absl::Floor(start, period);
     } else if (interval_distribution == "sync_burst_spread") {
       absl::Duration start = clock_->Now() - absl::UnixEpoch();
-      int nb_peers = peers_[action_state->rpc_service_index].size();
-      double fraction = (double)service_instance_ / nb_peers;
+      double nb_peers = peers_[action_state->rpc_service_index].size();
+      double fraction = service_instance_ / nb_peers;
       LOG(INFO) << "sync_burst_spread burst delay: " << fraction * period;
       action_state->next_iteration_time =
         period + absl::UnixEpoch() + absl::Floor(start, period) +
